+++ src/js/Autocomplete.js
@@ -1,0 +1,12 @@
+/*!
+MolView v2.2 (http://molview.org)
+Copyright (c) 2014, Herman Bergwerf
+ALL RIGHTS RESERVED
+*/
+"use strict";
+
+var Autocomplete = {
+	init: function()
+	{
+	},
+}